+++ src/include/enums.h
@@ -42,6 +42,7 @@ v8::Local<v8::Object> policy();
 v8::Local<v8::Object> predicates();
 v8::Local<v8::Object> retry_policy_values();
 v8::Local<v8::Object> status();
+v8::Local<v8::Object> subcode();
 v8::Local<v8::Object> ttl_enum_values();
 v8::Local<v8::Object> batchTypes();
 v8::Local<v8::Object> queryDuration();